Add property to control the behaviour of geoclue
This was required to implement the lat/lon parameters properly.
Previously Emerillon automatically updated the map position at
application start based on data received from Geoclue. To disable
this a property called "auto-update" has been added, with a default
value of FALSE, which is set in main.c according to the requirements.
Also this enables automatic position updating when moving (via gypsy/gpsd
Geoclue provider or similar) because we can install a signal handler.